tests/submission/Intel/audio/audio_muted_volumechange.html

<!DOCTYPE html>
<html>
  <head>
    <title>Audio Test: audio_muted_volumechange</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-media-muted" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if audio.muted is changed that expecting a simple event named volumechange is fired at the media element" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <audio id="m">The user agent doesn't support media element.</audio>
    <script type="text/javascript">
        var media = document.getElementById("m");
        var name = document.getElementsByName("assert")[0].content;
        var t = async_test(name);

        function startTest() {
            t.step(function() {
                assert_true(true, "muted changeed");
            });
            t.done();
        }

        media.addEventListener("volumechange", startTest, false);
        media.muted = !media.defaultMuted;
    </script>
  </body>
</html>
